+++ hutool-core/src/main/java/cn/hutool/core/date/Month.java
@@ -5,7 +5,8 @@ import java.util.Calendar;
 /**
  * 月份枚举<br>
  * 与Calendar中的月份int值对应
- * 
+ *
+ * @author Looly
  * @see Calendar#JANUARY
  * @see Calendar#FEBRUARY
  * @see Calendar#MARCH
@@ -19,53 +20,106 @@ import java.util.Calendar;
  * @see Calendar#NOVEMBER
  * @see Calendar#DECEMBER
  * @see Calendar#UNDECIMBER
- * 
- * @author Looly
- *
  */
 public enum Month {
-
-	/** 一月 */
+	/**
+	 * 一月
+	 */
 	JANUARY(Calendar.JANUARY),
-	/** 二月 */
+	/**
+	 * 二月
+	 */
 	FEBRUARY(Calendar.FEBRUARY),
-	/** 三月 */
+	/**
+	 * 三月
+	 */
 	MARCH(Calendar.MARCH),
-	/** 四月 */
+	/**
+	 * 四月
+	 */
 	APRIL(Calendar.APRIL),
-	/** 五月 */
+	/**
+	 * 五月
+	 */
 	MAY(Calendar.MAY),
-	/** 六月 */
+	/**
+	 * 六月
+	 */
 	JUNE(Calendar.JUNE),
-	/** 七月 */
+	/**
+	 * 七月
+	 */
 	JULY(Calendar.JULY),
-	/** 八月 */
+	/**
+	 * 八月
+	 */
 	AUGUST(Calendar.AUGUST),
-	/** 九月 */
+	/**
+	 * 九月
+	 */
 	SEPTEMBER(Calendar.SEPTEMBER),
-	/** 十月 */
+	/**
+	 * 十月
+	 */
 	OCTOBER(Calendar.OCTOBER),
-	/** 十一月 */
+	/**
+	 * 十一月
+	 */
 	NOVEMBER(Calendar.NOVEMBER),
-	/** 十二月 */
+	/**
+	 * 十二月
+	 */
 	DECEMBER(Calendar.DECEMBER),
-	/** 十三月，仅用于农历 */
+	/**
+	 * 十三月，仅用于农历
+	 */
 	UNDECIMBER(Calendar.UNDECIMBER);
 
 	// ---------------------------------------------------------------
+	private static final int[] DAYS_OF_MONTH = {31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31, -1};
+
+	/**
+	 * 对应值，见{@link Calendar}
+	 */
 	private final int value;
 
+	/**
+	 * 构造
+	 *
+	 * @param value 对应值，见{@link Calendar}
+	 */
 	Month(int value) {
 		this.value = value;
 	}
 
+	/**
+	 * 获取{@link Calendar}中的对应值
+	 *
+	 * @return {@link Calendar}中的对应值
+	 */
 	public int getValue() {
 		return this.value;
 	}
 
 	/**
+	 * 获取此月份最后一天的值，不支持的月份（例如UNDECIMBER）返回-1
+	 *
+	 * @param isLeapYear 是否闰年
+	 * @return 此月份最后一天的值
+	 */
+	public int getLastDay(boolean isLeapYear) {
+		int lastDay = DAYS_OF_MONTH[value];
+		if (isLeapYear && Calendar.FEBRUARY == value){
+			lastDay += 1;
+		}
+		return lastDay;
+	}
+
+	/**
 	 * 将 {@link Calendar}月份相关值转换为Month枚举对象<br>
-	 * 
+	 *
+	 * @param calendarMonthIntValue Calendar中关于Month的int值
+	 * @return {@link Month}
 	 * @see Calendar#JANUARY
 	 * @see Calendar#FEBRUARY
 	 * @see Calendar#MARCH
@@ -79,9 +133,6 @@ public enum Month {
 	 * @see Calendar#NOVEMBER
 	 * @see Calendar#DECEMBER
 	 * @see Calendar#UNDECIMBER
-	 * 
-	 * @param calendarMonthIntValue Calendar中关于Month的int值
-	 * @return {@link Month}
 	 */
 	public static Month of(int calendarMonthIntValue) {
 		switch (calendarMonthIntValue) {